requirements.txt
asgiref==3.8.1
cachetools==5.5.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-rest-authtoken==2.1.4
django-user-agents==0.4.0
djangorestframework==3.15.2
phonenumbers==8.13.55
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyYAML==6.0.2
rsa==4.9
sqlparse==0.5.3
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
